feat(estatisticas): Cria contrato de estatisticas

+++ src/main/java/com/florido/expense_tracker_api/models/Transacao.java
@@ -26,7 +26,7 @@ public class Transacao {
     private UUID id;
 
     @Column(name = "valor")
-    private long valor;
+    private double valor;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "tipo")
